+++ online-shopping-website/backend/src/routes/orders.ts
@@ -1,17 +1,28 @@
-import express, { Response, Request } from "express";
-import convertStringToArray from "../helpers/convertStringToArray";
-import isObjectOnlyInts from "../helpers/isObjectOnlyInts";
-import hasRequiredOrderCreationParams from "../helpers/verifyOrderCreation";
-import { allOrders, createOrder, deleteOrder, orderByItem, updateOrder } from "../prismaFunctions/orderFuncs";
+import express, { Response, Request } from 'express';
+import convertStringToArray from '../helpers/convertStringToArray';
+import isObjectOnlyInts from '../helpers/isObjectOnlyInts';
+import hasRequiredOrderCreationParams from '../helpers/verifyOrderCreation';
+import {
+  allOrders,
+  createOrder,
+  deleteOrder,
+  orderByItem,
+  updateOrder,
+  orderById,
+} from '../prismaFunctions/orderFuncs';
+import { objectFromRequest } from '../helpers/jwtFuncs';
+import { UserRole, User } from '@prisma/client';
 
 const orderRouter = express.Router();
 
-orderRouter.post("/create", async (req: Request, res: Response) => {
-  // TODO: make sure only customers are allowed to create orders
-
+orderRouter.post('/create', async (req: Request, res: Response) => {
+  const user = objectFromRequest(req);
   try {
+    if (user == undefined || user == null || (user as User).role !== UserRole.CUSTOMER) {
+      throw new Error('Invalid Authorization');
+    }
     const { itemIds, itemQuantities, err } = validateOrder(req, false);
-    if (err !== "") {
+    if (err !== '') {
       throw err;
     }
 
@@ -27,12 +38,24 @@ orderRouter.post("/create", async (req: Request, res: Response) => {
   }
 });
 
-orderRouter.delete("/delete", async (req: Request, res: Response) => {
-  const orderId = parseInt(req.query["id"] as string);
+orderRouter.delete('/delete', async (req: Request, res: Response) => {
+  const user = objectFromRequest(req);
+  const orderId = parseInt(req.query['id'] as string);
   try {
-    if (orderId === undefined || isNaN(orderId)) {
-      throw new Error("ID is invalid");
+    if (user == undefined || user == null) {
+      throw new Error(`Invalid authentication`);
     }
+    if (orderId === undefined || isNaN(orderId)) {
+      throw new Error('ID is invalid');
+    }
+    const oldOrder = await orderById({ orderId: orderId });
+    if (oldOrder == null) {
+      throw new Error('Order not found');
+    }
+    if (oldOrder.userId !== (user as User).id || (user as User).role !== UserRole.ADMIN) {
+      throw new Error('Invalid authentication');
+    }
+
     const deletedOrder = await deleteOrder({ orderId: orderId });
     res.json(deletedOrder).status(200);
   } catch (e) {
@@ -40,12 +63,23 @@ orderRouter.delete("/delete", async (req: Request, res: Response) => {
   }
 });
 
-orderRouter.post("/update", async (req: Request, res: Response) => {
+orderRouter.post('/update', async (req: Request, res: Response) => {
   // needs all items and quantities for it to work
-  const orderId = parseInt(req.query["id"] as string);
+  const user = objectFromRequest(req);
+  const orderId = parseInt(req.query['id'] as string);
   try {
+    if (user == undefined || user == null) {
+      throw new Error(`Invalid authentication`);
+    }
     if (orderId === undefined || isNaN(orderId)) {
-      throw new Error("ID is invalid");
+      throw new Error('ID is invalid');
+    }
+    const oldOrder = await orderById({ orderId: orderId });
+    if (oldOrder == null) {
+      throw new Error('Order not found');
+    }
+    if (oldOrder.userId !== (user as User).id) {
+      throw new Error('Invalid authentication');
     }
     const { itemIds, itemQuantities, err } = validateOrder(req, true);
     console.log(err);
@@ -65,20 +99,20 @@ orderRouter.post("/update", async (req: Request, res: Response) => {
   }
 });
 
-orderRouter.get("/find", async (req: Request, res: Response) => {
-  const search = parseInt(req.query["id"] as string);
+orderRouter.get('/find', async (req: Request, res: Response) => {
+  const search = parseInt(req.query['id'] as string);
   try {
     if (search === undefined || isNaN(search)) {
-      throw new Error("ID is invalid");
+      throw new Error('ID is invalid');
     }
     const orders = await orderByItem({ itemId: search });
-    res.json({ orders }).status(200);
+    res.json(orders).status(200);
   } catch (e) {
     res.status(400).json({ error: e, message: e.meta?.cause || e.message });
   }
 });
 
-orderRouter.get("/all", async (req: Request, res: Response) => {
+orderRouter.get('/all', async (req: Request, res: Response) => {
   await allOrders()
     .then((orders) => {
       res.status(200).json(orders);
@@ -94,27 +128,27 @@ export default orderRouter;
 function validateOrder(req: Request, update: boolean) {
   let itemIds: number[] = [];
   let itemQuantities: number[] = [];
-  let err = "";
-  if (typeof req.body.itemIds === "string") itemIds = convertStringToArray(req.body.itemIds);
-  else if (typeof req.body.itemIds === "object" && isObjectOnlyInts(req.body.itemIds))
+  let err = '';
+  if (typeof req.body.itemIds === 'string') itemIds = convertStringToArray(req.body.itemIds);
+  else if (typeof req.body.itemIds === 'object' && isObjectOnlyInts(req.body.itemIds))
     itemIds = req.body.itemIds as number[];
   else {
     itemIds = [];
-    err = err.concat("Invalid Item Ids ");
+    err = err.concat('Invalid Item Ids ');
   }
 
-  if (typeof req.body.itemQuantities === "string") itemQuantities = convertStringToArray(req.body.itemQuantities);
-  else if (typeof req.body.itemQuantities === "object" && isObjectOnlyInts(req.body.itemQuantities))
+  if (typeof req.body.itemQuantities === 'string') itemQuantities = convertStringToArray(req.body.itemQuantities);
+  else if (typeof req.body.itemQuantities === 'object' && isObjectOnlyInts(req.body.itemQuantities))
     itemQuantities = req.body.itemQuantities as number[];
   else {
     itemQuantities = [];
-    err = err.concat("Invalid Item Quantities ");
+    err = err.concat('Invalid Item Quantities ');
   }
 
   if (itemIds.length !== itemQuantities.length) {
     itemIds = [];
     itemQuantities = [];
-    err = err.concat("Invalid arrays ");
+    err = err.concat('Invalid arrays ');
   }
 
   if (
@@ -126,6 +160,6 @@ function validateOrder(req: Request, update: boolean) {
     }) &&
     !update
   )
-    err = err.concat("Invalid Data ");
+    err = err.concat('Invalid Data ');
   return { itemIds: itemIds, itemQuantities: itemQuantities, err: err };
 }
